chore: rename project and package
