+++ pom.xml
@@ -7,12 +7,12 @@
     <version>3.5.1</version>
   </parent>
   
-  <groupId>com.example</groupId>
-  <artifactId>helloworld-agent</artifactId>
+  <groupId>akka.haiku</groupId>
+  <artifactId>akka-haiku</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 
-  <name>helloworld-agent</name>
+  <name>akka-haiku</name>
 
   <dependencies>
     <!-- Your dependencies go here -->
