+++ src/main/java/akka/haiku/domain/TextInputCollectorEvent.java
@@ -1,4 +1,4 @@
-package com.example.domain;
+package akka.haiku.domain;
 
 import java.util.List;
